Exams/Exam-02/Exam-02-1/Code/MoneyBankAPI/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
mysql-connector-python
aiomysql
python-dotenv
pytest
httpx
mysqlclient